Add "debug" builtin
Will show some information, currently just prints the `Repl` instance.

Also refactor checking if a command exists to its own function. Fixed a
bug in which we only checked for the command name, not with a slash.
This lead to some false positives, i.e. we would "find" the command
"bar", if a command called "foobar" exists in the $PATH.