Janky first version of history navigation (^P/^N)
Currently the history navigation is not quite correct, the order is off.
This version is very naïve, and the logic is a bit off. But it's
something, as a PoC!

Starting to feel more and more like a real shell!

...until I realize that we don't have piping, redirection, any
expansion/globbing (except for ~), and the visual bugs. More fun to
come! :D