break: Implement lexing and AST
We now lex the input and then parse the lexed tokens into an Abstract
Syntax Tree. This allows for much more sophisticated input parsing.

The AST includes support for basic expansions, such as tilde, globbing,
parameter and command. Only tilde expansion is actually implemented at
the moment.

The old `builtins` approach has been removed, it wasn't very nice to
begin with. This means some (temporary) removals, such as the `history`
command. Only `cd` and `exit` are currently implemented.

NOTE: This commit is huge because it was under development for about a
month, and the code base is so young that many changes were required
throughout it to allow for this.