feat: Add support for Engine contained in memory
We had halfway support for this before, but it never worked. Now, the
execution is implemented on each generic `Engine`, making it possible to
write the output to a `Vec<u8>` if in memory, otherwise as intended.

This can be used to verify that our parsing/execution works as expected,
as part of a regression suite. Although it doesn't test the actual
`Engine<std::io::Stdout>` that's used in the REPL, it should be good
enough for now.

Rework `History` to be a trait, with two implementations: the previous
is implemented as `FileHistory`, and a new `DummyHistory` which doesn't
do anything at all – which is used in the in-memory engine.