fix: Ghost quotes when incomplete
In the REPL, when a user would enter a single or double quote, the
closing one would be printed as well. This was kind of cool, but
ultimately doesn't really fit the vision.

Now, the quoted string tokens contain the state of whether or not they
were finished. If unfinished, we highlight them with another color.