break: Remove `tokenize` and add parameter to `lex`
Previously we had a `lex` function, used to generate tokens based on an
input string, as well as a `tokenize`. `lex` was simply a specialization
of `tokenize`, with the flag to ignore space always set. No point in
having two functions doing pretty much the same, only causing confusion
on which to use.