feat: Remove internal testing of engine, add test script
The internal engine testing was very wonky, requiring lots of time to
figure out how to make the engine work for different 'backends'. I.e.
both the interactive version (using `Stdout`), and an in-memory for
tests (using `Vec<u8>`). I finally had enough and removed it, replacing
it with a bash script that builds and runs the debug release.