misc: Clean up input cursor move to start position
At the end of most key event reads, we'd move the cursor to the position
for the next command. This meant many places of duplicate code, and in
one case, even some unnecessarily complex logic. These have all been
replaced with a function which is called at the end of the segments.