fix: Expand variables that do not exist into nothing
Previously, we would only try to expand a variable if it was known to
the environment. I.e. `echo $HOME` would (most times) expand, while
`echo $foo` would print the literal `$foo`.

Fix this by checking if the variable did not get expanded into a value,
and if so, expand it to nothing.