Implement tokenizer
This uses the `Consumer<char>` trait to tokenize a stream of characters.
Probably missing some tokens currently.

This tokenizer is completely semantically unaware, and almost all tokens
are single characters. The next step would be to parse the tokens
semantically. Not sure if it should be part of the AST, or another step
before that.