feat: Add semantic tokenizing
Name pending, especially the file name.

The semantic tokenizer takes a stream of "raw" tokens as input, and
transforms them into a more semantically aware version. Quotes and
escapes are kept, but the words are aware of them, i.e. they do not
break if we're in a quoted or escaped state.

Currently, command substitution and arithmetic expressions are not
supported (nor are subshells). I'm having the toughest time trying to
figure out how to handle these in the best way, they're the reason I
keep getting stuck at the parsing stage. Grrr