feat: Skeleton for AST parsing
Create the data structures I think we'll need (parsed from the POSIX
shell specification).

Also create the `Parser` trait, which consumes a
stream of `SemanticToken`s in order to generate the syntax tree. This is
still very much WIP, and currently the only parsers implemented are for
variable assignment and redirections.

Now I'm leaving to eat a steak and have a beer. Happy Saturday!