feat: remove v2 module + keep whitespace in AST
N.B.: this commit is a squash of around 40 commits, containing many many
changes (and the squash being made a long time after initially writing
them). This commit message probably does not do it justice, but I do not
have the patience to update it entirely. Just be aware that the
following list is _very_ inconclusive of the actual commit. Side note:
need to figure out a better flow when working on a feature branch for so
long.

So many changes here. Two big ones are:

- Remove the v2 module, i.e. replace the old parser with the new one.
  This is what makes the commit not work, since the callers are not
  updated correctly yet.
- Rework the AST to keep track of whitespace, in order to be able to
  reconstruct the command line entered by the user, from the AST itself.
  This was a way bigger change than I initially thought, requiring
  making `SimpleCommand` way uglier. The reason being that we need to
  keep the order of the assignments/redirections/words. Thus, the
  immensely horrible `SimpleCommandMeta` enum was born. Gotta find a
  better solution to this.