BROKEN: Move ToString implementations to module
Create the `ast::reconstruct` module which contains all the ToString
implementations. This is an attempt to make the `ast` module more
readable, and in the future it might be split into multiple files.

I never see people having their tests in another file. But they're
always in a `<module>::tests` module, so it should be easy to move out.
I'll do that. :-)