fix: Avoid std::process::exit as it bypasses constructors
Doing ^d on an empty line means exit; this changes the implementation
from exiting the process immediately to just emitting a "exit" token for
the repl to execute.

Doing a std::process::exit to exit the process will bypass any Rust
deconstructors defined. Currently, the posh repl uses a deconstructor to
reset the terminal from raw mode. If this isn't run, the terminal is
left in a an unusable state until a reset is triggered.