feat: Restructure the AST according to spec
The AST was somewhat in line with the official shell grammar rules
before, but now we're aiming to follow it much more closely.

A lot of the parsing is not there yet, and the same for the syntax
highlighting. With this commit, not much has actually functionally
changed, but it's a step in an interesting direction.