feat: Implement --tokenize/--lex/--ast for the REPL
If posh is run with either of these flags without specifying a --command
or a file, the shell will not execute the commands entered into the
REPL. Instead, the tokens/semantic tokens/AST will get printed.

The one exception is if the input is exactly "exit", which will exit the
shell. Note that any leading or trailing whitespace, or quotes, will not
exit.