feat: Basic signal handling
Change the `ExitStatus` struct to an enum, either containing a code or a
signal. In reality, a signal is also a code, but there is a semantic
difference between the two. If a process gets stopped by a signal, we
show the signal name in the prompt.

Apparently, signal handling in Rust has quite a history. Handling the
SIGINT signal appears to be different from other signals. Thus, we pull
in the `ctrlc` crate, setting up a handler that just tells the prompt to
display "SIGINT".