feat: Make CompleteCommand into an enum
Previously, the definition of CompleteCommand was a bit awkward. It had
two optional fields, one for a `List` (and optionally a `SeparatorOp`),
and one for a `Comment`. The problem was that the two should never be
able to be `None` at the same time.

Fix this by turning the struct into an enum. Now, it's either a `List`
(optionally with a `SeparatorOp` and/or a `Comment`), or just a
`Comment`.