feat: ParseResult type for AST node parsers
From the beginning, I knew I would one day have to replace the `Option`s
returned by the AST node parsers, with something that enables more
context. Surprisingly to me, the reason that made me switch was not
contextual error messages (although this is a step in that direction as
well), but rather to allow the parsers to say that "hey, you're on the
right track with this line, but I need more before it's valid". Thus,
the `ParseResult` and `ParseError` types was born.

Unfortunately, this solution makes the parser quite a bit uglier – no
longer can we simply propagate `None` cases, we need to handle specific
error cases manually. This is an area I want to improve upon in the
future, but this branch has lived long enough and provides such value to
the program that I feel comfortable with the state it's in currently.

With this patch, unfinished quotes (e.g. `'foo` or simply `\`) are
handled in the REPL as well. Syntax highlighting currently does not work
for multiline inputs, it treats each line as its own AST.

Some other minor changes in this commit (I would have loved to split it
up, but as I've mentioned in previous "mega" commits, I lack the
discipline...) include:

- Fixed bug where <> redirection would append
- Expand quotes before computing existence of program
- Niche: respect escaping state when searching for quotes in a string
  (Word::find)