fix: Highlighting issues for multiline commands
When a command spans multiple lines, i.e. when the PS2 prompt (what is
this actually called? hm) was invoked, the highlighting of commands got
weird. Examples are unfinished quotes, a pipeline with newlines after
the pipe, etc. I can't really explain exactly _how_ they were weird, but
it was often wrong – and being on the last line in the terminal buffer
behaved differently than all the others.

With this fix, multiline commands seem to behave quite well in most
circumstances. It is likely far from perfect, and a lot of edge cases
are bound to pop up – I'm starting to dislike the entire `input` module
more and more... probably time to figure out how to use an actual
readline implementation while still having syntax highlighting.

Unfortunately, this solution pretty much nullifies the previous commit.
The solution I opted for currently means having two "different"
attributes for the content of a word: one with all "\\\n" instances
removed (the main one), and one with them kept. This is ugly, but
something like this is needed to let the input module know when it
should break the lines.

Also implement `syntax_highlighting::Highlighter` for `Word`. The only
reason this wasn't done earlier was because "words should have different
colors depending on the context"... my brother in Christ, the color can
be set before the `Word::write_highlighted` invocation. :-D