fix: (more) Correctly handle incomplete pipe sequences
The first command in a pipe sequence was handled differently compared to
the rest of them, with regards to incompleteness. If the first command
was incomplete, we stored it in the pipe sequence and returned it as
incomplete. But if a trailing command was incomplete, it would get
discarded.

With this fix, all commands in a pipe sequence are treated equivalently
with regards to incompleteness. :-)