fix: Parameter expansion by pipeline
When executing e.g. `a=b; echo $a`, `b` should be printed. Previously,
expansion happened on the entire AST. This made little sense, as some
parts should be expanded before others. I don't think this solution is
correct, nor is it very nice, as we expand pipeline by pipeline where in
reality (I think) it should be word by word. But I believe that
pipelines are currently incorrectly implemented, so in order to get
something that resembles the expected behavior, we handle pipelines with
one command differently from one with many – namely, a command with only
assignments will _not_ update the assignment list if part of a
multi-command pipeline.

Also remove `Expand` implementations for nodes above `Pipeline`, as it
is incorrect to expand on such high levels.