feat: New file-based test.sh
Replace the old test.sh with a file-based one. This runs all files in
the test/ directory (except for any found in test/expected/) and
compares the stdout and stderr against the same file in
test/expected/{stdout,stderr}. If a corresponding expected file does not
exist, the test is not run. The new pattern should allow for far more
readable tests. :-)

This also removes the dependency towards bash! We used bash's literal
newline feature in the previous version of the script, but no longer.
Update the Makefile to use `sh` rather than `bash` to run the test
suite.