misc: Move builtins to their own module
I should say move _back_, since we had a similar thing long ago. During
one of the initial major rewrites of the parser (perhaps when the AST
was first introduced?), it was removed and moved to functions on the
`Engine` to have something quick and working.

Having their own module has many advantages I think, but the biggest one
is the discoverability. All files (except for mod.rs) in the builtin
directory should be a single builtin, making it easy to find what we
support and their source code. :-)

Still not 100% sure if the implementation is perfect, i.e. having the
same names in the functions as the file. While I ponder if this can be
nicer, we have a starting point. :-)