feat: Make abbreviations expand again
Abbreviations is a "REPL-only" feature, inspired by fish's feature of
the same name. It's like an alias, but it expands in real time while in
an interactive environment. This makes it (perhaps) a bit more intuitive
when compared to traditional aliases, and can make your history file
clearer.

Abbreviations somewhat worked before, but they wouldn't visually expand.
This was also lost during a previous major rewrite. This new
implementation is quite naive, and only allows abbreviations at the
start of a line. I'm sure there are some visual (and functional) issues
to sort out still, but it seems to work OK for my usual use cases.

With this patch, the `Highlighter::write_highlighted` API has changed:
it now takes a `Context` object, currently containing the starting
column (as used for multiline inputs), as well as a flag to say if
abbreviations should be highlighted or not. :-)