misc: Move builtin execution to builtin module
Instead of checking if the supplied command is a valid builtin in the
engine, a new `execute` function has been implemented in the builtin
module. This does the same logic, but feels a bit better since we now
only have to update builtin/mod.rs when implementing a new builtin. :-)