break: Make Engine non-generic + streamline builtins
Previously, the `Engine` struct could be made from anything
that implements `std::io::Write`. The idea was that, for instance, one
could initiate an Engine that prints to a vector buffer instead of
stdout, which would be quite handy for tests. A long time ago I wrestled
with this, and could never come up with a good solution. And the
genericness has just been left behind since then. Well, no more! I
officially give up on that idea, but I may return to it some day as I
still believe it has potential. :-)

One commit, two changes! The builtin module has been streamlined a bit.
All builtins now expose a function named `execute` with the signature
`&mut Engine, &[&str] -> Result<ExitStatus>`. The module also gained a
new `get` function for retrieving the appropriate builtin from a name.
This removes the need to update two functions when implementing a new
builtin, which is of course more DRY :-)