feat: Expand $PS1 and $PS2 + separate env from assignments
When reading $PS1 and $PS2 for the prompt, we now turn them into a
`Word` and expand them before printing. This allows you to have e.g.
`PS1='$PWD$ '`. Cool!

Previously, we added all environment variables into the Engine's
`assignments` field for every CompleteCommand execution. This suddenly
felt very clunky and weird to me, so with this patch, we don't do that
anymore. This also fixed an issue in which the prompt would only update
for the next command, before turning back into its old value. That also
brings me to a cool thing with this patch: you can actually set the $PS1
and $PS2 variables from within the shell session and have it update
immediately. Cool again! :-)