misc: Move all input reading to input module
Previously, the logic for when to keep reading lines (i.e. when the AST
is incomplete), was done in the REPL. Now, the input modules exposes a
single function: `read_full_command(&mut Engine) -> Result<String>`,
that encapsulates this. A bit cleaner API, I think :-)

And one more thing: I usually like checking a shell's performance by
just holding enter and seeing how it behaves. Very scientific, I know.
In most cases, it's clean, but in some cases, artifacts can be seen. I
recently noticed this was the case in psh... so I tried to rectify it in
this patch as well, by only moving to the first column when enter has
been pressed, if the inputted line is non-empty. Better, but I think I
can still see some color weirdness. To be improved!