feat: Rudimentary $? expansion
$? is a special parameter according to the spec. When invoked, it shall
expand to the last executed pipeline's exit code. If none exists, I
_think_ it should be 0, since that seems to be how other shells behave.

With this patch, the `last_status` field has been moved from the REPL to
the Engine. It's type has also been changed from
`Option<Vec<ExitStatus>>`, to `Vec<ExitStatus>`, and the default value
on entry is a single 0. For those unaware: the reason for the "last
status" to be a vector is that I may want to re-implement the feature in
which the exit code of each command in a pipeline gets shown in the
prompt (yet another fish feature!).

The `last_status` field gets updated for every pipeline execution.

The actual implementation isn't very nice, and it will for sure be
revisited in the future. But it's something! With every one of these
kinds of features, it feels more and more like a real shell... which
makes me very happy! :-)