fix: Escape double quote when double quoted
Escaped double quote characters within a double quoted string shall be
replaced by the literal double quote. I.e., `echo "\"foo\""` should
print `"foo"`. This was not the case before, but is now!

This doesn't seem like the correct solution to me. The spec says that
"the quote characters backslash, single quote, and double quote shall be
removed unless they themselves are quoted". But for this solution, if
inside a double quote and a backslash is encountered, whether or not it
gets removed depends on the next character (i.e. if it's a double
quote). Hmmmm