break: Rely on Word being expanded before execution
Instead of expanding a word when checking if it's a valid command,
expect it to be expanded before-hand. This makes sense since the amount
of expansions should be fewer this way, and the APIs get simpler and
more easy to use.

Also move the `expand_alias` function to the `Engine` struct, and inline
the recursion into the function itself.