feat: `builtins` builtin
Very meta! The `builtins` builtin prints all available builtins, or if
supplied with an argument, exits with 0 if the argument is a builtin,
otherwise 1.

Minor refactor in the `builtin` module as well, we now store all
builtins in a `const BUILTINS: &[(&str, Builtin)]`, allowing us to query
it from other places. Maybe not too surprising, but this is how the
`builtins` builtin gets its data :-)