feat: Rework execution logic
Add the "nix" crate as a dependency, and start doing external command
execution via the classic fork/exec rather than having an awkward
function returning a process that hasn't been spawned yet? It was really
bad before, this makes it less bad. Hopefully it should also make it
easier to implement features in the future, especially ones depending on
subshells (which I imagine we'll be able to implement via fork as well).

A bugfix part of this patch is that builtins are now able to be
redirected and piped.

A regression: alias expansion does not really work with this patch. One
of the main reasons is that I want to make the implementation better in
the future, as currently (or rather, before) the target of an alias
could only ever be a simple command. It should (probably) be allowed to
put an entire SyntaxTree there, so what we want to do is something along
the lines of replacing all aliases that are the first words in simple
commands in the input string, then re-parse with that. TBD! Also left a
comment in the `expand_alias` function regarding this.