fix: Syntax highlighting for unfinished lists
We would only tell the parser that a `List` was unfinished if the first
`AndOrList` was, not if any of the others. Fix this!

This is the same type of fix that was needed for e.g. `Pipeline`s (or
rather `PipeSequence`s) which was done in an earlier commit. This one
slipped through the chairs, so to say. There may be others like it still
lurking... :-)