misc: Make invalid command color less striking
The color for an invalid command was a bit harsh, and at the same time,
the color for invalid syntax ("unparsed") code was a bit dull. Kill two
birds by simply swapping these colors. :-)

This may be an acceptable solution to #6. To be honest, I thought I had
done this a long time ago but evidently not. Sorry gevhaz!