fix: Crash on "incomplete" lists
In 4f3e217259e7a2cd056cb4f11a70a9579d935a06, a major regression slipped
through, in which executing e.g. `ls;` would make the parser think that
there's an unfinished AndOrList after the separator. The issue is that
when parsing a pipeline, two things can make it be "unfinished": a
command in the pipeline being unfinished, or the non-first command
simply not existing (e.g. as in `ls |`). The latter is interesting,
since it needs to be handled outside of the "normal" unfinished
propagation. It turns out, we need to do the exact same thing for
AndOrLists! Quite obvious in hindsight, as most things are. :-)