misc: Minor readability improvements
When parsing redirections, use the `Option::zip` method to make the
match arms become a bit more readable.

In the function checking if a character is a valid part of a "Name", use
the `char::is_ascii_alphanumeric` method instead of matching on the
specific alphanumerics.

And a simple join for equivalent, non-capturing patterns in the match
part of the function checking if a string is a valid "Name".