perf: Replace ToString with AsRef<str> for ReservedWord
Previously, `ReservedWord` implemented `ToString`, which allocates a new
`String` on the heap. There was no fancy logic, string concatenation, or
anything of the sort in the ToString implementation, so we replace it
with `AsRef<str>` instead. This does not allocate any memory. Very minor
optimization, but it's something I noticed during some other problem
solving and wanted to make sure I fix it right now.

This is also the first commit with the `perf:` prefix! The idea is that
any change that is made solely for performance gain, or otherwise
space/time optimization. I think it's a good idea, let's see how it
turns out. :-)