feat: Recursive AST parsing in words (cmd sub) + more
NOTE: no command substitution is actually performed with this patch,
this is only internal parsing work. But it does make command
substitutions behave correctly in syntax highlighting!

This commit has been brewing for a while. I'm still lacking the
discipline, and as I now realize probably also the git know-how, to
separate this into the (probably quite a few) commits it deserves.
During my development work, I made _some_ half decent commit messages,
which I'm attaching to the end of this commit message for reference.

---

Much has changed in this commit, I don't think I'll be able to go into
details about all of it. I'll try my best to describe the main problem
it solves.

Previously, the `semtok` tokenizer had logic for figuring out when a
"Word" ended. This means quoting, backslash escaping, and even an
attempt at figuring out when we were in a command substitution
environment or arithmetic expression, and in what level. This was always
incorrect, ugly, and other negative adjectives. Instead, this is now all
done by the parser; the semtok module is quite a bit dumber, but it
allows the parser to recurse when a command substitution start (`$(`) is
encountered, by invoking the main parse method again. In order to figure
out when to stop, we check if the unparsed section of the command
substitution starts with a `)`, meaning that it's time to say "all
finished, let's continue on the layer above". This makes it easier to
deal with, grants more readable code, and perhaps most importantly for
the interactive shell: allows for correct syntax highlighting when
inside a command substitution. (note: the backtick variant of command
substitution is not at all supported yet)

With this patch, the `tok` and `semtok` modules are awfully similar,
bringing me a step closer to a goal I've had for quite some time now: to
remove one of them. Actually, the main reason for having both was that I
found it easier to deal with when I thought I had to find the start and
end of a "word" (as described above). Now I neither need that or even
want it, and it shows – since they have become so similar. One beautiful
day I will remove one.

In order to propagate errors from a command substitution, I also made
the switch to make the `Parser::parse` method have the return type of
`std::result::Result<SyntaxTree, std::result::Result<SyntaxTree,
ParseError>`. The `Ok(SyntaxTree)` case is a finished, good AST. The
`Err(Ok(SyntaxTree))` case is an unfinished syntax tree, while the
`Err(Err(ParseError))` case means that a syntax (or otherwise
unrecoverable) error occurred during parsing. I feel like this is both
kind of hideous, and kind of elegant at the same time. Though it also
makes me wish that the standard library defined a type alias like `type
Either<A, B> = Result<A, B>;`, just because the signature of the parse
method would become shorter. :-)

There are a few parts of the code base that I've been unhappy with
before, and have learned to be even unhappier with during this task. The
major one has to do with `Word`s. The fact that it contains two
variables, one with escaped newlines (pretty much only needed for syntax
highlighting/interactive use), and one without. This is unmaintainable,
ugly, and wasteful, since they are almost equivalent and are both
heap-allocated strings. Also, there's a bug with word expansion, in
which after quote removal, the known expansions' ranges don't get
updated. All of the `is_finished` logic I think I want to remove, in
favor of having that in the parser itself, and perhaps as a single
`finished: bool` field in the Word struct. Still thinking about it!

Even though this is by far my longest commit message to date (ever), I
know I'm probably missing quite a bit. But this was the reason for the
change, some implementation details, and some rant-y parts about how
unhappy I am with parts of the code base still. But to leave on a
positive note: this patch feels reeeally cool to use, seeing the syntax
highlighting in command substitution makes me very giddy. I've been
dreading the word recognition part for a very long time, and it has been
holding me back from other feature development since I want to have it
in a good, "finished" state. This brings us a lot closer to that! :-)

---

[1] WIP: parse word multiple times

In order to support a word being e.g. `$(cmd)hello$var`, we need to keep
parsing for acceptable SemanticTokens until none are left.

This rewrites parts of both Token and SemanticToken, to "dumb them
down", i.e. just keep the dollar symbol as a single token rather than
consuming them as part of a word.

[2] WIP: fix highlighting index offset

[3] WIP: looking better!

Previously, a `SemanticToken::Word` would contain single quotes, double
quotes, dollar signs, and equals signs. These have been removed, and are
now getting parsed in the AST parser instead. I think it makes more
sense and allows for finer control when parsing the input. It's also a
big step in the direction of completely removing the `Token` part of the
tokenizing, so hopefully soon we can rename the `SemanticToken` to
`Token` and skip the first part entirely. Because right now, Token and
SemanticToken are veeery similar.

[4] fix: Expand parameters into correct variable

Previously, variables were expanded on the `name` field of a Word. This
left the `name_with_escaped_newlines` field not being expanded. Instead,
do the expansions on the latter, and re-calculate the former based on
that.

Really need to figure out a better solution than having two fields. I
think the initial idea was to be backwards compatible, but I'm starting
to second guess that decision. What if we have a single `name` field,
that contains the escaped newlines, and then have a method for returning
it with those removed? Maybe that could work, will try it out in the
future. More pressing things now though :-)

[5] WIP: ugly single quote solution

ugly ugly ugly

[6] WIP: remove quoted string parsers from semtok