break: Newtype pattern for LeadingWhitespace
Previously, LeadingWhitespace was a simple type alias to String. I think
it makes a little more sense to have it be its own type, simply a
wrapper over a String instead. This makes it possible to implement
`serde::Serialize` for it specifically.

It is a bit uglier, having to use the `.0` syntax in some places, but I
think the pros outweigh this con. :-)