feat: Better serialization + add serde feature
The serde serialization has been improved a bit, by manually
implementing the `Serialize` trait for many structs/enums. Also remove
the serde_json dependency on psh-core, since it's only used by psh.

New features for both psh and psh-core, both called "serde". Without
this on psh-core, the dependency to the serde crate is not active, and
the `serialize` module will not be active. For psh,
the `serde_json` dependency will not be active, so when running with
`--ast`, it will print the SyntaxTree's Debug implementation.

There are still quite a few structs deriving Serialize, which is a bit
awkward, requiring lots of `#[cfg_attr(feature = "serde",
derive(Serialize)]`. Not sure if this is idiomatic, or if there's a way
to do derivations in another module, or if we should manually implement
all of them.