feat: --json flag to psh, propagate serde dependency
If the "serde" feature is enabled for psh, it will depend on the "serde"
feature of psh-core.

New `--json` flag, only available with the "serde" feature. With this,
when using the `--ast` flag, JSON output will be printed instead of the
debug implementation of SyntaxTree. We had JSON output specifically when
running `psh -c <cmd> --ast` before, i.e. not with a file or
interactively. Now it's available anywhere based on this flag. :-)