break: Remove "duplicate" field on Word
When parsing a word, a backslash followed by a newline character should
be removed and regarded as line continuation (if unquoted or in double
quotes). However, we need to store these newline characters for syntax
highlighting to work on multiline inputs. This put us in a bit of a
conundrum, since after expansion (quote removal), how do we know if the
user actually entered an empty string, or if it was just a series of 1
or more escaped newlines? The solution we had was to store two different
fields, one with the escaped newlines removed and one with them kept. I
knew since the beginning that this was a bad idea, but I really wanted
to keep making progress so I opted for it with the intention of fixing
it in the future.

I think I finally came up with a good solution: always store these
escaped newlines in the word, and when expanding a SimpleCommand, remove
any suffixes that only contained escaped newlines. This isn't perfect,
since it will only work on the suffixes in a command (i.e. not the main
word, and not in other places where it probably matters). But there's a
rework of the expansion coming soon where I will hopefully fix this. For
know, less code is better code, and this just feels better. Stay tuned
:-)