feat: More on-spec PS1 expansion + \w
The prompt variable PS1 should not be expanded exactly like a regular
word, but it should expand parameters and command substitutions (and a
unique ! expansion). Introduce a new `expand_prompt` function to do
this, but no command substitution expansion yet.

Also, off-spec but a nice QoL thing: expand \w into tilde compressed
$PWD. There will likely be more, similar ones in the future, such as
hostname. \u for user may not be implemented at all, since there are
environment variables for it.