feat: Implement `ast` builtin
This may replace the `--ast` flag (and will in that case likely be
followed by a `tokenize` builtin), if we can deem this to cover all
areas that the flag does. I'm mainly thinking about quoting and such,
i.e. is `psh -c 'ast ls ;'` always able to do the same as
`psh --ast 'ls ;'`? With more complex commands, of course.

The `ast` builtin supports a `-j`/`--json` flag if the serde feature is
enabled, in which case the output will be printed in JSON format.
Otherwise, it will print the debug implementation of the `SyntaxTree`.
Also unsure about this.. maybe the serde feature should always print in
JSON format, i.e. no flags? Need to think about this. :-)