misc: Remove History trait
Previously, we implemented this trait for both the "normal"
`FileHistory`, and a `DummyHistory` which was supposed to be used with
an `Engine` that wrote to a vector (for testing) – this was back when
`Engine` was generic.

Also remove the iterator implementation, it wasn't even used. May return
in the future though!