misc: Rework History fn signatures a bit
The `read` function previously returned `Result<Option<&String>>`, where
a mistake had slipped through: if the cursor was out of bounds,
`Ok(None)` would be returned. Then we would effectively check this again
and return `Err(Error::HistoryOutOfBounds)`. Simply remove the `Option`
part of the signature, and return the error if out of bounds.