misc: Remove unused code, refactor config, general housekeeping
We had quite a few functions that were never used, tried to remove all
of them.

The config module has been removed and replaced with a color module.
Instead of having static functions on a `Colors` ZST, have the functions
at the top level of this module.

Some general housekeeping as well, sorting `derive`s on the AST nodes
alphabetically and such.