feat: Increment $SHLVL upon REPL start
When starting the REPL, increment the SHLVL environment variable. If it
is unset, set it to 1.

We read the variable as a single byte. I am of the thinking that if you
have more than a byte's worth of nested shell environments, you have
bigger issues and shouldn't care if the variable stops incrementing. :-)