misc: rustfmt now supports let-else!
The day has come! ...probably quite a few days ago, but I haven't been
working on the project in a while. And before that, I didn't update my
podman environment for an even longer while.

But nevertheless, rustfmt now supports let-else blocks! Which means we
need to fix them. It was becoming a game for me, instead of following
news regarding this I would update rustfmt and running `make lint` in
this repo to see if it would be caught. And it finally was :-) Great job
to the rustfmt team!