test: Fix wrongful assumption about $HOME
Previously we would expect the running user's $HOME entry to be located
at /home/<username>. In most cases this is true, but in some it isn't:
most obviously, root usually has /root as $HOME. And my podman dev
environment has $HOME set to /cfg. So instead of having `/home/user` in
the expected stdout, and replacing "user" with "$USER", we now have a
"<workdir>" entry in the expected, and replace that with "$HOME".

Side note: `podman run` would set the home entry in /etc/passwd to the
value passed to `--workdir`. This was confusing the heck out of me,
since `echo ~` and `echo ~<user>` would give two different values! I
realized after some research that there's a `--passwd-entry` option to
`podman run`, which I now make use of. Good stuff :-)